requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
jinja2==3.1.5
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
weasyprint>=66.0
